feat(#15): lotusCard 폰트, 너비 스타일 적용하기
- 폴더 구조 변경하기
- LotusCard 폰트, 너비 스타일 적용하기
