+++ .vscode/settings.json
@@ -6,5 +6,5 @@
   ],
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": "always"
-  }
+  },
 }
