+++ apps/frontend/src/feature/Lotus/index.tsx
@@ -1,0 +1,18 @@
+import {
+  LotusAuthor,
+  LotusCreateDate,
+  LotusDescription,
+  LotusLink,
+  LotusLogo,
+  LotusProvider,
+  LotusTitle
+} from './Lotus';
+
+export const Lotus = Object.assign(LotusProvider, {
+  Title: LotusTitle,
+  Description: LotusDescription,
+  Author: LotusAuthor,
+  Logo: LotusLogo,
+  CreateDate: LotusCreateDate,
+  Link: LotusLink
+});
