+++ apps/frontend/src/page/lotus.lazy.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@froxy/design/components';
 import { createLazyFileRoute } from '@tanstack/react-router';
-import { LotusCard } from '@/widget/LotusCard';
+import { LotusCardList } from '@/widget/LotusList/LotusCardList';
 
 export const Route = createLazyFileRoute('/lotus')({
   component: RouteComponent
@@ -11,7 +11,7 @@ function RouteComponent() {
     <div>
       <Button onClick={() => alert('Lotus')}>Lotus</Button>
       <h1 className="text-red-300">Lotus world</h1>
-      <LotusCard />
+      <LotusCardList />
     </div>
   );
 }
